Factored common values for wrappers
