+++ var/spack/repos/builtin/packages/openmpi/package.py
@@ -2,10 +2,10 @@ import os
 
 from spack import *
 
-from spack.provider_contracts import MpiProvider
+from spack.provider_contracts import StandardMpiProvider
 
 
-class Openmpi(MpiProvider, Package):
+class Openmpi(StandardMpiProvider, Package):
     """Open MPI is a project combining technologies and resources from
        several other projects (FT-MPI, LA-MPI, LAM/MPI, and PACX-MPI)
        in order to build the best MPI library available. A completely
@@ -39,22 +39,6 @@ class Openmpi(MpiProvider, Package):
     provides('mpi@:3.0', when='@1.7.5:')
 
     depends_on('hwloc')
-
-    @property
-    def cc_compiler_wrapper(self):
-        return join_path(self.prefix.bin, 'mpicc')
-
-    @property
-    def cxx_compiler_wrapper(self):
-        return join_path(self.prefix.bin, 'mpicxx')
-
-    @property
-    def f77_compiler_wrapper(self):
-        return join_path(self.prefix.bin, 'mpif77')
-
-    @property
-    def fc_compiler_wrapper(self):
-        return join_path(self.prefix.bin, 'mpifort')
 
     def url_for_version(self, version):
         return "http://www.open-mpi.org/software/ompi/v%s/downloads/openmpi-%s.tar.bz2" % (version.up_to(2), version)
